Move controllers under pkg folder
Now all operator's modules, except "controllers", are located in
"pkg" folder.

To keep it unified, this commit puts "controllers" in the same
place.
