+++ cmd/cluster-cloud-controller-manager-operator/main.go
@@ -35,7 +35,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/healthz"
 
 	configv1 "github.com/openshift/api/config/v1"
-	"github.com/openshift/cluster-cloud-controller-manager-operator/controllers"
+	"github.com/openshift/cluster-cloud-controller-manager-operator/pkg/controllers"
 	// +kubebuilder:scaffold:imports
 )
