+++ pkg/controllers/cloud_config_sync_controller_test.go
